pyproject.toml
[project]
name = "feeder-model-training"
version = "0.1.0"
description = "Training scripts for feeder bee detection models (POLO and localizer)"
requires-python = ">=3.10"
dependencies = [
    "mosaic-behavior[polo,localizer] @ git+https://github.com/ecodylicscience/mosaic.git",
]

[project.optional-dependencies]
dev = ["ipython"]
